Refactoring the project !
