+++ src/Components/TodoList/index.jsx
@@ -1,7 +1,6 @@
-import { Check } from "react-bootstrap-icons";
-import { XLg } from "react-bootstrap-icons";
-import Options from "../Options/Options";
 import { useState } from "react";
+import { Check, XLg } from "react-bootstrap-icons";
+import Options from "./Options";
 
 function TodoList({ todo, setTodo }) {
   const [isCompleted, setIsCompleted] = useState(false);
@@ -18,7 +17,7 @@ function TodoList({ todo, setTodo }) {
       // } else {
       //   console.log("no");
       // }
-      isCompleted ? setIsCompleted(false) : setIsCompleted(true)
+      isCompleted ? setIsCompleted(false) : setIsCompleted(true);
     }
   }
 
@@ -31,11 +30,7 @@ function TodoList({ todo, setTodo }) {
   return (
     <div className="bg-white dark:bg-input-dark rounded-lg shadow-lg shadow-light-shadow dark:shadow-dark-shadow">
       {todo.map((myTodo, index) => (
-        <div
-          key={index}
-          data-index={index}
-          className="todo py-5 border-b-2 dark:border-text-dark my-todo"
-        >
+        <div key={index} data-index={index} className="todo py-5 border-b-2 dark:border-text-dark my-todo">
           <div className="px-3 flex justify-between items-center">
             <div className="flex items-center">
               <Check
@@ -43,11 +38,7 @@ function TodoList({ todo, setTodo }) {
                 className={`check-btn ${isCompleted ? "completed" : null}`}
                 onClick={(e) => markAsCompleted(e, index)}
               />
-              <p
-                className={`todo-content ${
-                  isCompleted ? "line-through" : null
-                }`}
-              >
+              <p className={`todo-content ${isCompleted ? "line-through" : null}`}>
                 {myTodo.charAt(0).toUpperCase() + myTodo.slice(1)}
               </p>
             </div>
